refactor: extract scheduler helpers
